fix : 데이터 타입 불일치 수정

+++ server/src/main/java/com/pathfinder/server/diary/dto/DiaryDto.java
@@ -39,7 +39,7 @@ public class DiaryDto {
         private String area1;
         private String area2;
         private Long recommendedCount;
-        private Long scrapCount;
+        private Integer scrapCount;
         private Integer views;
         private LocalDateTime createdAt;
         private LocalDateTime modifiedAt;
